motor-triagem-ia/requirements.txt
fastapi>=0.109.0
sqlalchemy>=2.0.0
pgvector>=0.2.4
psycopg2-binary>=2.9.9
httpx>=0.26.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
openai>=1.0.0
uvicorn[standard]>=0.27.0
